requirements-dev.txt
beautifulsoup4>=4.14.3
aiohttp>=3.13.3
homeassistant>=2026.4.2
pytest>=9.0
pytest-asyncio>=1.3.0
pytest-homeassistant-custom-component>=0.13.323
aioresponses>=0.7.8
ruff>=0.15
mypy>=1.15
python-dotenv>=1.2
